requirements.txt
pandas
pypyodbc
bcrypt
django-pyodbc